🔧 fix: Separate forgot-password into server/client components to prevent static generation with useSignIn hook
