+++ app/(auth)/forgot-password/page.tsx
@@ -1,144 +1,8 @@
-"use client";
-
 // Force dynamic rendering to prevent static generation during build
 export const dynamic = 'force-dynamic';
 
-import { useState } from "react";
-import { useSignIn } from "@clerk/nextjs";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import * as z from "zod";
-import { Loader2 } from "lucide-react";
-import { toast } from "sonner";
-
-import { Button } from "@/components/ui/button";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
-import Link from "next/link";
-
-// Define form validation schema
-const formSchema = z.object({
-  email: z.string().email({ message: "Please enter a valid email address" }),
-});
-
-type FormValues = z.infer<typeof formSchema>;
+import { ForgotPasswordClient } from './client';
 
 export default function ForgotPasswordPage() {
-  const { isLoaded, signIn } = useSignIn();
-  const [isLoading, setIsLoading] = useState(false);
-  const [emailSent, setEmailSent] = useState(false);
-
-  // Initialize form with react-hook-form
-  const form = useForm<FormValues>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      email: "",
-    },
-  });
-
-  // Handle form submission
-  async function onSubmit(values: FormValues) {
-    if (!isLoaded) return;
-
-    try {
-      setIsLoading(true);
-      
-      // Start the password reset process
-      await signIn.create({
-        strategy: "reset_password_email_code",
-        identifier: values.email,
-      });
-      
-      setEmailSent(true);
-      toast.success("Password reset email sent. Please check your inbox.");
-      
-    } catch (err: unknown) {
-      console.error("Password reset error:", err);
-      const errorMessage = err && typeof err === 'object' && 'errors' in err && Array.isArray((err as { errors: unknown[] }).errors) && (err as { errors: { message?: string }[] }).errors[0]?.message
-        ? (err as { errors: { message: string }[] }).errors[0].message
-        : "Failed to send reset email. Please try again.";
-      toast.error(errorMessage);
-    } finally {
-      setIsLoading(false);
-    }
-  }
-
-  return (
-    <div className="flex min-h-screen items-center justify-center p-4">
-      <Card className="w-full max-w-md">
-        <CardHeader>
-          <CardTitle className="text-2xl">Reset Password</CardTitle>
-          <CardDescription>
-            {emailSent 
-              ? "Check your email for a reset link" 
-              : "Enter your email to receive a password reset link"}
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          {emailSent ? (
-            <div className="text-center space-y-4">
-              <p>
-                We&apos;ve sent password reset instructions to your email.
-                Please check your inbox and follow the link to reset your password.
-              </p>
-              <p className="text-sm text-muted-foreground">
-                If you don&apos;t see the email, check your spam folder.
-              </p>
-            </div>
-          ) : (
-            <Form {...form}>
-              <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-                <FormField
-                  control={form.control}
-                  name="email"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Email</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="email"
-                          placeholder="name@example.com"
-                          autoComplete="email"
-                          disabled={isLoading}
-                          {...field}
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-
-                <Button type="submit" className="w-full" disabled={isLoading}>
-                  {isLoading ? (
-                    <>
-                      <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                      Sending...
-                    </>
-                  ) : (
-                    "Send Reset Link"
-                  )}
-                </Button>
-              </form>
-            </Form>
-          )}
-        </CardContent>
-        <CardFooter className="flex justify-center border-t p-4">
-          <p className="text-sm text-muted-foreground">
-            Remember your password?{" "}
-            <Link href="/sign-in" className="text-blue-600 hover:text-blue-800">
-              Sign in
-            </Link>
-          </p>
-        </CardFooter>
-      </Card>
-    </div>
-  );
+  return <ForgotPasswordClient />;
 }
